requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro-datasets[pandas-csvdataset, pandas-exceldataset, pandas-parquetdataset, plotly-plotlydataset, plotly-jsondataset, matplotlib-matplotlibwriter]>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.12
notebook
scikit-learn~=1.5.1
seaborn~=0.12.1
tensorflow>=2.12.0
holidays
pandas~=2.1.3
matplotlib~=3.7.1
numpy~=1.26.4
uvicorn
fastapi
streamlit
plotly
requests
python-dotenv
azure-storage-blob
streamlit_option_menu